requirements.txt
#使用したパッケージ

pillow == 5.2.0
tweepy == 3.7.0